JAMES-3007 HealthCheck components equals, hashCode
