+++ core/src/test/java/org/apache/james/core/healthcheck/ComponentNameTest.java
@@ -19,40 +19,15 @@
 
 package org.apache.james.core.healthcheck;
 
-import java.util.Objects;
+import org.junit.jupiter.api.Test;
 
-import com.google.common.base.MoreObjects;
+import nl.jqno.equalsverifier.EqualsVerifier;
 
-public class ComponentName {
-    private final String name;
+class ComponentNameTest {
 
-    public ComponentName(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    @Override
-    public final boolean equals(Object o) {
-        if (o instanceof ComponentName) {
-            ComponentName that = (ComponentName) o;
-
-            return Objects.equals(this.name, that.name);
-        }
-        return false;
-    }
-
-    @Override
-    public final int hashCode() {
-        return Objects.hash(name);
-    }
-
-    @Override
-    public String toString() {
-        return MoreObjects.toStringHelper(this)
-            .add("name", name)
-            .toString();
+    @Test
+    void shouldMatchBeanContract() {
+        EqualsVerifier.forClass(ComponentName.class)
+            .verify();
     }
 }
